Address a few long-standing compiler warnings.
git-svn-id: https://svn.filezilla-project.org/svn/FileZilla3/trunk@9022 ce1b9ea0-41a8-11dc-9b12-cb80f3b22977

+++ src/engine/ftp/list.h
@@ -45,7 +45,7 @@ private:
 	bool viewHidden_{}; // Uses LIST -a command
 
 	// Listing index for list_mdtm
-	int mdtm_index_{};
+	size_t mdtm_index_{};
 
 	fz::monotonic_clock time_before_locking_;
 };
